Added prework files
Welcome to the dynamic world of web development! In this portfolio project, I've embarked on an exciting journey to create a sophisticated banking application using the powerful Django framework. My main goal was to guide myself through each step of development, honing my skills while crafting an app loaded with features that showcase my abilities as a capable developer.